+++ agents/sirji_agents/researcher/researcher.py
@@ -13,6 +13,7 @@ from sirji_messages import message_parse, MessageFactory, ActionEnum
 
 from .embeddings.factory import EmbeddingsFactory
 from .inferer.factory import InfererFactory
+from ..decorators import retry_on_exception
 
 
 DEFAULT_SKIP_LIST = [
@@ -110,7 +111,7 @@ class ResearchAgent:
                 "OpenAI API key is not set as an environment variable")
 
         # Initialize OpenAI client
-        self.client = OpenAI(api_key=api_key)
+        self.client = OpenAI(api_key=api_key, timeout=60)
 
     def message(self, input_message):
         """Public method to process input messages and dispatch actions."""
@@ -416,6 +417,7 @@ class ResearchAgent:
         self.logger.info("Indexing processed files")
         return self._embeddings_manager.upload_batches(self.file_streams)
         
+    @retry_on_exception()
     def create_assistant(self, body):
         self.logger.info("Creating a new assistant instance")
         """
@@ -433,7 +435,7 @@ class ResearchAgent:
             return
 
         # Initialize OpenAI client
-        client = OpenAI(api_key=api_key)
+        client = OpenAI(api_key=api_key, timeout=60)
 
         project_folder = os.environ.get("SIRJI_PROJECT")