+++ sirji/vscode-extension/src/py_scripts/sync_file.py
@@ -50,10 +50,15 @@ class SyncFile:
 
     sirji_installation_dir = os.environ.get("SIRJI_INSTALLATION_DIR")
    
-    agent_id = "RESEARCHER"
-    installed_agent_folder = os.path.join(sirji_installation_dir, 'studio', 'agents')
-    reasearcher_config_path = os.path.join(installed_agent_folder, f'{agent_id}.yml')
-    config_file_contents = self.read_file(reasearcher_config_path)
+    default_path = os.path.join(os.path.dirname(__file__), '../defaults/agents/RESEARCHER.yml')
+    agent_path = os.path.join(sirji_installation_dir, 'studio', 'agents', 'RESEARCHER.yml')
+
+    if os.path.exists(agent_path):
+        agent_config_path = agent_path
+    else:
+        agent_config_path = default_path
+
+    config_file_contents = self.read_file(agent_config_path)
     config = yaml.safe_load(config_file_contents)
 
     llm = config['llm']